+++ src/Avatar3D.cpp
@@ -343,49 +343,3 @@ inline Eye::Eye(double feature_5, bool left)
     m_draw_lines = true;
 }
 
-bool SwarmOfAvatars3D::mousePressEvent(QMouseEvent* event)
-{
-    qWarning() << "(Swarm) mouse pressed";
-    return true;
-}
-
-bool SwarmOfAvatars3D::mouseReleaseEvent(QMouseEvent* event)
-{
-    qWarning() << "(Swarm) mouse released";
-    return true;
-}
-
-bool SwarmOfAvatars3D::mouseMoveEvent(QMouseEvent* event)
-{
-    qWarning() << "(Swarm) mouse moved";
-    return true;
-}
-
-bool SwarmOfAvatars3D::wheelEvent(QWheelEvent* event)
-{
-    qWarning() << "(Swarm) mouse wheel event";
-    return true;
-}
-
-#include <QtWidgets>
-
-inline QWidget* SwarmOfAvatars3D::create_prop_widget() {
-    QWidget* widget = new QWidget();
-    QVBoxLayout* layout = new QVBoxLayout(widget);
-
-    layout->addWidget(new QLabel(QString::fromUtf8("MIEJSCE NA TWOJĄ REKLAMĘ")));
-
-    widget->resize(layout->sizeHint());
-    widget->setMinimumSize(layout->sizeHint());
-    widget->setMaximumSize(layout->sizeHint());
-
-    widget->setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Preferred);
-
-    return widget;
-}
-
-//inline QVector<PropWidget*> SwarmOfAvatars3D::create_propwidget_and_get_subwidgets()
-//{
-//    return PropSwarmOfAvatars3D::create_and_get_subwidgets(this);
-//}
-
